Add Apm info
This commit adds some basic  APM integration - it injects a transaction into WorkerLoop around "abortableCompute".
This transaction is named with the pipelin.id, and enables the flow of events through the filters and outputs to be tracked
by creating child spans around each of the filters and outputs.
This commit uses the automatic attach method, and requires an `elasticapm.properties` file to be defined - by default this
will be read from `src/main/resources`, but the location can be changed via a system property